third_party/blink/web_tests/external/wpt/css/css-position/backdrop-ua-style-expected.txt

This is a testharness.js-based test.
[FAIL] UA styles for ::backdrop
  assert_equals: expected "fixed" but got "absolute"
Harness: the test ran to completion.
